remove review mappings
